pyproject.toml
[project]
name = "downly"
version = "0.1.0"
description = "touka bot for downloading media from various sources, including youtube, twitter, and more."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx==0.25.1",
    "loguru>=0.7.3",
    "msgpack>=1.1.1",
    "pika>=1.3.2",
    "psutil>=7.0.0",
    "psycopg2-binary==2.9.9",
    "pydantic>=2.11.7",
    "pyrogram==2.0.106",
    "pyyaml==6.0.1",
    "sqlalchemy==2.0.23",
    "sqlalchemy-utils==0.41.1",
    "tgcrypto==1.2.5",
    "uvloop==0.19.0",
    "yt-dlp==2023.7.6",
]
